refactor: restructure components into a parent register directory
