+++ components/Register/TabShortR.tsx
